requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
cryptography==46.0.2
distlib==0.4.0
Django==5.2.7
django-allauth==65.11.2
django-allauth-suap==0.1.2
django-crispy-forms==2.4
django-filter==25.2
django-htmx==1.26.0
django-simple-menu==2.1.4
django-tables2==2.7.5
django-tinymce==4.1.0
filelock==3.19.1
identify==2.6.15
idna==3.10
nodeenv==1.9.1
oauthlib==3.3.1
platformdirs==4.4.0
pre_commit==4.3.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
ruff==0.13.3
sqlparse==0.5.3
urllib3==2.5.0
virtualenv==20.34.0
